latest stable

+++ FrontEnd/src/app/components/buyer/order.js
@@ -3,6 +3,7 @@ import { connect } from 'react-redux';
 import config from './../../../app-config';
 import { getSections, getItems } from './../../redux/actions/menu-actions';
 import { setCartCount } from './../../redux/actions/order-actions';
+import { setMessage } from './../../redux/actions/util-action';
 
 class BuyerOrder extends React.Component { 
     constructor(props) {
@@ -39,7 +40,7 @@ class BuyerOrder extends React.Component {
         if(this.state.display_items) {
             array.push(
                 <div className='g-order-rest-title'>
-                   <img src={config.base  + this.props.selectedOrderRest.image}/>
+                   <img src={config.base  + this.props.selectedOrderRest.image} alt="NO DISPLAY"/>
                    <div>
                        <span style={{fontSize: '38px'}}> {this.props.selectedOrderRest.rest_name} </span>  
                        <span className='g-secondary-text'> Address </span>
@@ -60,7 +61,7 @@ class BuyerOrder extends React.Component {
                         array.push(
                             <div className="g-menu-row">
                                 <div className="g-menu-image">
-                                     <img className="g-image" src={config.base + item.image} alt="NO PHOTO"/>
+                                     <img className="g-image" src={config.base + item.image} alt="NO DISPLAY"/>
                                 </div>
                                 <div className="g-menu-desc">
                                     <div className="g-menu-desc-title">{item.name}</div>
@@ -85,7 +86,7 @@ class BuyerOrder extends React.Component {
             });
             if(this.state.total > 0) {
                 array.push(
-                    <div className="g-section-head" style={{textAlign: "end" , paddingRight: "68px"}}>
+                    <div className="g-section-head" style={{justifyContent: "flex-end" , paddingRight: "68px"}}>
                         <button className="btn btn-success"  style={{marginRight: "24px"}} 
                         onClick={() => this.addToCart()}> ADD TO CART</button>
                         <span style={{width: "120px"}}>TOTAL: ${this.state.total}</span>
@@ -93,7 +94,7 @@ class BuyerOrder extends React.Component {
                 )
             } else {
                 array.push(
-                    <div className="g-section-head" style={{textAlign: "end", paddingRight: "68px"}}>
+                    <div className="g-section-head" style={{justifyContent: "flex-end", paddingRight: "68px"}}>
                         <span style={{width: "120px"}}>TOTAL: ${this.state.total}</span>
                     </div> 
                 )
@@ -102,23 +103,40 @@ class BuyerOrder extends React.Component {
         return array;
     }
     addToCart() {
-        let itemsAdded = {
-            rest_id: this.props.selectedOrderRest.id,
-            total: this.state.total,
-            items: [],
-            cust_id: JSON.parse(localStorage.getItem('user1')).id
-        };
-        Object.keys(this.state.display_items).forEach(key => { 
-            if(this.state.display_items[key].length > 0) {
-                this.state.display_items[key].forEach(item => { 
-                    if(item.count > 0) {
-                        itemsAdded.items.push(item)
-                    }
+        let confirm = true;
+        if( JSON.parse(localStorage.getItem('cart')) !== null) {
+            confirm = window.confirm('You have items in your cart from a different order. Are you sure you want to replace them with this one?');
+        }
+        if(confirm) {
+            let itemsAdded = {
+                rest_id: this.props.selectedOrderRest.id,
+                total: this.state.total,
+                items: [],
+                cust_id: JSON.parse(localStorage.getItem('user1')).id
+            };
+            Object.keys(this.state.display_items).forEach(key => { 
+                if(this.state.display_items[key].length > 0) {
+                    this.state.display_items[key].forEach(item => { 
+                        if(item.count > 0) {
+                            itemsAdded.items.push(item)
+                        }
+                    })
+                } 
+            })
+            try {
+                localStorage.setItem('cart', JSON.stringify(itemsAdded));
+                this.props.setCartCount();
+                this.props.setMessage({
+                    msg: 'Successfully added items to your cart',
+                    name: 'success'
                 })
-            } 
-        })
-        localStorage.setItem('cart', JSON.stringify(itemsAdded));
-        this.props.setCartCount();
+            } catch {
+                this.props.setMessage({
+                    msg: 'Could not add items to your cart',
+                    name: 'danger'
+                })
+            }
+        }
     }
     decCount(val) {
         let sec,id;
@@ -176,7 +194,8 @@ const mapDispatchToProps = dispatch => {
     return {
         getSections: payload => dispatch(getSections(payload)),
         getItems: payload => dispatch(getItems(payload)),
-        setCartCount: () => dispatch(setCartCount())
+        setCartCount: () => dispatch(setCartCount()),
+        setMessage: payload => dispatch(setMessage(payload))
     };
 }
 export default connect(mapStateToProps, mapDispatchToProps)(BuyerOrder);
